gobrightbox no longer returns pointers to slices
Boy, using a typed language sure does make this kind of change easy!

If this was written in Ruby I'd have missed several of these and I'd be
getting bug reports for weeks (depending on the quality of my tests ;)

Though if this was written in Ruby the whole project would be finished
by now.